Symbian support for project. New UNUSABLE_MMAP code path for platforms that cannot use the required flags (Xbox 360, Symbian). Falls back to malloc() instead. N1 define for compilers (GCCE) that don't like to assume (int)-1 will safely cast to (char)-1. Miscellaneous space->tab fixes.

+++ Core/HLE/sceKernelThread.cpp
@@ -822,7 +822,7 @@ void sceKernelCheckThreadStack()
 {
   u32 error;
   Thread *t = kernelObjects.Get<Thread>(__KernelGetCurThread(), error);
-  u32 diff = (u32)abs((s64)t->stackBlock - (s64)currentMIPS->r[MIPS_REG_SP]);
+  u32 diff = abs((long)((s64)t->stackBlock - (s64)currentMIPS->r[MIPS_REG_SP]));
   ERROR_LOG(HLE, "%i=sceKernelCheckThreadStack()", diff);
 	RETURN(diff); //Blatant lie
 }
